Add new classes for sensor data integration
This commit introduces several new classes to handle the integration of sensor data. A ConfigManager class is implemented for managing configuration data, loaded from a given YAML configuration file. CommonSensorDataIntegrationService class serves to send the sensor data to a designated URL, including a built-in retry mechanism. This commit also includes additions of data logging request class and reading dataclass files for Sentek sensors data. All these changes have been made to aid in the integration of various sensors' data and to enhance modularity and maintainability of the code.