Add device registration feature and update sensor data handling
This commit introduces a new feature that allows registration of sensor devices. In addition to this, the sensor data classes were transformed into Data Transfer Objects (DTOs), which conform with standard design principles and practices. This move to DTOs enables the clear separation of the data layer from the rest of the application and refines encapsulation. This update covers the Agvolution, Sentek, and Weenat services. The sensor data handling process has also been improved to make the application more robust and efficient. Test files were also renamed and moved appropriately to match with new changes. Lastly, minor corrections were made in the README file as well. The goal of these changes is to enhance the structure of the application and improve maintainability.