Change API URL to use HTTPS in config.yaml
The API url has been updated to use the HTTPS protocol instead of HTTP for added security and data protection. This change was due to the need for secure data transmission when interacting with the API from different applications. HTTPS ensures that the data is encrypted before transmission, dramatically reducing the risk of data breaches or malicious activities.